+++ juriscraper/opinions/united_states/state/delaware.py
@@ -6,6 +6,7 @@ Reviewer: mlr
 """
 
 import re
+from urllib.parse import urljoin
 
 from juriscraper.OpinionSiteLinear import OpinionSiteLinear
 
@@ -14,7 +15,7 @@ class Site(OpinionSiteLinear):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.url = (
-            "http://courts.delaware.gov/opinions/list.aspx?ag=supreme%20court"
+            "https://courts.delaware.gov/opinions/List.aspx?ag=supreme%20court"
         )
         # Note that we can't do the usual thing here because 'del' is a Python keyword.
         self.court_id = "juriscraper.opinions.united_states.state.del"
@@ -26,16 +27,22 @@ class Site(OpinionSiteLinear):
 
         :return: None
         """
-        for row in self.html.xpath("//table/tr"):
-            self.cases.append(
-                {
-                    "name": row.xpath("./td[1]/a/text()")[0].strip(),
-                    "date": row.xpath("./td[2]/text()")[0].strip(),
-                    "docket": row.xpath("./td[3]/a/text()")[0].strip(),
-                    "url": row.xpath("./td[1]/a/@href")[0].strip(),
-                    "judge": " ".join(row.xpath("./td[6]/text()")),
-                }
-            )
+        for row in self.html.xpath("//table//tr[not(th)]"):
+            case = {
+                "name": row.xpath("td[1]/a/span/text()")[0].strip(),
+                "date": row.xpath("td[2]/text()")[0].strip(),
+                "docket": row.xpath("td[3]/a/span/text()")[0].strip(),
+                "url": urljoin(
+                    self.url, row.xpath("td[1]/a/@href")[0].strip()
+                ),
+                "judge": " ".join(row.xpath("td[6]/text()")).strip(),
+                "per_curiam": False,
+            }
+            if "per curiam" in case["judge"].lower():
+                case["judge"] = ""
+                case["per_curiam"] = True
+
+            self.cases.append(case)
 
     def extract_from_text(self, scraped_text: str) -> dict:
         """Extract lower court from the scraped text.